Collection of TEP-Scripts
Particle Analysis is the Master script for the analysis of single stain images image Analysis output data. The script requires the Helper Script TEP-Functions to run, TEP_graphics_plugin is optional and is designed for plotting the particle size distribution on the individual filters.
Particle-Analysis-TSP is a slightly adjusted version (still requiring TEP-Functions) that can be controlled via the script TSP-Master. This is the go to solution, since it takes all three output files from analysing the double stained images and processing them together. A tydier version of this will be uploaded by the end of the year.